Location inject supports parent location from name
The current functionality of the Locations import is not able to define a new location by specifying its parent as name ( string ).

Usecase: Importing the locations, it is possible to define the "parent". But the import of this value was "as is"
this means, no lookup for the locations ID has been done. If the ID had been passed, it was ok, otherwise no value added.